Apps that utilise the Hasheous API can now be issued keys to allow them to use the Proxy service (#65)
* Rebuilt dataobject building to use a server side template

* Added support for storing app screenshots

* Screenshots are now zoomable

* WIP

* Completed Client API key

+++ hasheous/Controllers/V1.0/DataObjectController.cs
@@ -1,8 +1,12 @@
 using System.Data;
+using Authentication;
 using Classes;
 using gaseous_signature_parser.models.RomSignatureObject;
+using hasheous_server.Classes;
 using hasheous_server.Models;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Identity.UI.Services;
 using Microsoft.AspNetCore.Mvc;
 
 namespace hasheous_server.Controllers.v1_0
@@ -14,6 +18,43 @@ namespace hasheous_server.Controllers.v1_0
     [Authorize]
     public class DataObjectsController : ControllerBase
     {
+        private readonly UserManager<ApplicationUser> _userManager;
+        private readonly SignInManager<ApplicationUser> _signInManager;
+        private readonly IEmailSender _emailSender;
+        private readonly ILogger _logger;
+
+        public DataObjectsController(
+            UserManager<ApplicationUser> userManager,
+            SignInManager<ApplicationUser> signInManager,
+            IEmailSender emailSender,
+            ILoggerFactory loggerFactory)
+        {
+            _userManager = userManager;
+            _signInManager = signInManager;
+            _emailSender = emailSender;
+            _logger = loggerFactory.CreateLogger<AccountController>();
+        }
+
+        [MapToApiVersion("1.0")]
+        [HttpGet]
+        [ProducesResponseType(typeof(DataObjectDefinition), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [AllowAnonymous]
+        [Route("{ObjectType}/Definition")]
+        public async Task<IActionResult> DataObjectDefinition(Classes.DataObjects.DataObjectType ObjectType)
+        {
+            try
+            {
+                var DataObjectDefinition = hasheous_server.Classes.DataObjects.DataObjectDefinitions[ObjectType];
+
+                return Ok(DataObjectDefinition);
+            }
+            catch (Exception ex)
+            {
+                return NotFound();
+            }
+        }
+
         [MapToApiVersion("1.0")]
         [HttpGet]
         [ProducesResponseType(StatusCodes.Status200OK)]
@@ -44,6 +85,18 @@ namespace hasheous_server.Controllers.v1_0
             }
             else
             {
+                if (ObjectType == Classes.DataObjects.DataObjectType.App)
+                {
+                    var user = await _userManager.GetUserAsync(User);
+
+                    DataObjectPermission dataObjectPermission = new DataObjectPermission(_userManager);
+                    DataObject.Permissions = dataObjectPermission.GetObjectPermission(user, ObjectType, DataObject.Id);
+                    if (DataObject.Permissions.Contains(DataObjectPermission.PermissionType.Update))
+                    {
+                        DataObject.UserPermissions = dataObjectPermission.GetObjectPermissionList(DataObject.Id);
+                    }
+                }
+
                 return Ok(DataObject);
             }
         }
@@ -55,17 +108,28 @@ namespace hasheous_server.Controllers.v1_0
         [Route("{ObjectType}")]
         public async Task<IActionResult> NewDataObject(Classes.DataObjects.DataObjectType ObjectType, Models.DataObjectItemModel model)
         {
-            hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
+            // check permission
+            var user = await _userManager.GetUserAsync(User);
+            DataObjectPermission dataObjectPermission = new DataObjectPermission(_userManager);
 
-            Models.DataObjectItem? DataObject = DataObjects.NewDataObject(ObjectType, model);
-
-            if (DataObject == null)
+            if (dataObjectPermission.CheckAsync(user, ObjectType, DataObjectPermission.PermissionType.Create).Result)
             {
-                return NotFound();
+                hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
+
+                Models.DataObjectItem? DataObject = DataObjects.NewDataObject(ObjectType, model);
+
+                if (DataObject == null)
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    return Ok(DataObject);
+                }
             }
             else
             {
-                return Ok(DataObject);
+                return Unauthorized();
             }
         }
 
@@ -75,20 +139,31 @@ namespace hasheous_server.Controllers.v1_0
         [Route("{ObjectType}/{Id}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public async Task<IActionResult> EditDataObject(Classes.DataObjects.DataObjectType ObjectType, long Id)
+        public async Task<IActionResult> DeleteDataObject(Classes.DataObjects.DataObjectType ObjectType, long Id)
         {
-            hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
+            // check permission
+            var user = await _userManager.GetUserAsync(User);
+            DataObjectPermission dataObjectPermission = new DataObjectPermission(_userManager);
 
-            Models.DataObjectItem? DataObject = DataObjects.GetDataObject(ObjectType, Id);
-
-            if (DataObject == null)
+            if (dataObjectPermission.CheckAsync(user, ObjectType, DataObjectPermission.PermissionType.Delete, Id).Result)
             {
-                return NotFound();
+                hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
+
+                Models.DataObjectItem? DataObject = DataObjects.GetDataObject(ObjectType, Id);
+
+                if (DataObject == null)
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    DataObjects.DeleteDataObject(ObjectType, Id);
+                    return Ok();
+                }
             }
             else
             {
-                DataObjects.DeleteDataObject(ObjectType, Id);
-                return Ok();
+                return Unauthorized();
             }
         }
 
@@ -100,39 +175,69 @@ namespace hasheous_server.Controllers.v1_0
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> EditDataObject(Classes.DataObjects.DataObjectType ObjectType, long Id, Models.DataObjectItemModel model)
         {
-            hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
+            // check permission
+            var user = await _userManager.GetUserAsync(User);
+            DataObjectPermission dataObjectPermission = new DataObjectPermission(_userManager);
 
-            Models.DataObjectItem? DataObject = DataObjects.EditDataObject(ObjectType, Id, model);
-
-            if (DataObject == null)
+            if (dataObjectPermission.CheckAsync(user, ObjectType, DataObjectPermission.PermissionType.Update, Id).Result)
             {
-                return NotFound();
+                hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
+
+                Models.DataObjectItem? DataObject = DataObjects.EditDataObject(ObjectType, Id, model);
+
+                if (DataObject == null)
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    return Ok(DataObject);
+                }
             }
             else
             {
-                return Ok(DataObject);
+                return Unauthorized();
             }
         }
 
         [MapToApiVersion("1.0")]
         [HttpPut]
-        [Authorize(Roles = "Admin,Moderator")]
+        // [Authorize(Roles = "Admin,Moderator,Member")]
+        [Authorize]
         [Route("{ObjectType}/{Id}/FullObject")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> EditDataObject(Classes.DataObjects.DataObjectType ObjectType, long Id, Models.DataObjectItem model)
         {
-            hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
-
-            Models.DataObjectItem? DataObject = DataObjects.EditDataObject(ObjectType, Id, model);
-
-            if (DataObject == null)
+            if (ModelState.IsValid)
             {
-                return NotFound();
+                // check permission
+                var user = await _userManager.GetUserAsync(User);
+                DataObjectPermission dataObjectPermission = new DataObjectPermission(_userManager);
+
+                if (dataObjectPermission.CheckAsync(user, ObjectType, DataObjectPermission.PermissionType.Update, Id).Result)
+                {
+                    hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
+
+                    Models.DataObjectItem? DataObject = DataObjects.EditDataObject(ObjectType, Id, model);
+
+                    if (DataObject == null)
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        return Ok(DataObject);
+                    }
+                }
+                else
+                {
+                    return Unauthorized();
+                }
             }
             else
             {
-                return Ok(DataObject);
+                return BadRequest(ModelState);
             }
         }
 
@@ -164,19 +269,30 @@ namespace hasheous_server.Controllers.v1_0
         [Route("{ObjectType}/{Id}/Attributes")]
         public async Task<IActionResult> NewDataObjectAttribute(Classes.DataObjects.DataObjectType ObjectType, long Id, AttributeItem model)
         {
-            hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
+            // check permission
+            var user = await _userManager.GetUserAsync(User);
+            DataObjectPermission dataObjectPermission = new DataObjectPermission(_userManager);
 
-            Models.DataObjectItem? DataObject = DataObjects.GetDataObject(ObjectType, Id);
-
-            if (DataObject == null)
+            if (dataObjectPermission.CheckAsync(user, ObjectType, DataObjectPermission.PermissionType.Update, Id).Result)
             {
-                return NotFound();
+                hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
+
+                Models.DataObjectItem? DataObject = DataObjects.GetDataObject(ObjectType, Id);
+
+                if (DataObject == null)
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    AttributeItem attributeItem = DataObjects.AddAttribute(Id, model);
+
+                    return Ok(attributeItem);
+                }
             }
             else
             {
-                AttributeItem attributeItem = DataObjects.AddAttribute(Id, model);
-
-                return Ok(attributeItem);
+                return Unauthorized();
             }
         }
 
@@ -187,19 +303,30 @@ namespace hasheous_server.Controllers.v1_0
         [Route("{ObjectType}/{Id}/Attributes/{AttributeId}")]
         public async Task<IActionResult> DeleteDataObjectAttribute(Classes.DataObjects.DataObjectType ObjectType, long Id, long AttributeId)
         {
-            hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
+            // check permission
+            var user = await _userManager.GetUserAsync(User);
+            DataObjectPermission dataObjectPermission = new DataObjectPermission(_userManager);
 
-            Models.DataObjectItem? DataObject = DataObjects.GetDataObject(ObjectType, Id);
-
-            if (DataObject == null)
+            if (dataObjectPermission.CheckAsync(user, ObjectType, DataObjectPermission.PermissionType.Update, Id).Result)
             {
-                return NotFound();
+                hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
+
+                Models.DataObjectItem? DataObject = DataObjects.GetDataObject(ObjectType, Id);
+
+                if (DataObject == null)
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    DataObjects.DeleteAttribute(Id, AttributeId);
+
+                    return Ok();
+                }
             }
             else
             {
-                DataObjects.DeleteAttribute(Id, AttributeId);
-
-                return Ok();
+                return Unauthorized();
             }
         }
 
@@ -210,6 +337,12 @@ namespace hasheous_server.Controllers.v1_0
         [Route("{ObjectType}/{Id}/SignatureMap")]
         public async Task<IActionResult> GetDataObjectSignatureMap(Classes.DataObjects.DataObjectType ObjectType, long Id)
         {
+            // signatures aren't valid for apps
+            if (ObjectType != Classes.DataObjects.DataObjectType.App)
+            {
+                return NotFound();
+            }
+
             hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
 
             Models.DataObjectItem? DataObject = DataObjects.GetDataObject(ObjectType, Id);
@@ -231,6 +364,12 @@ namespace hasheous_server.Controllers.v1_0
         [Route("{ObjectType}/{Id}/SignatureMap")]
         public async Task<IActionResult> NewDataObjectSignatureMap(Classes.DataObjects.DataObjectType ObjectType, long Id, long SignatureId)
         {
+            // signatures aren't valid for apps
+            if (ObjectType != Classes.DataObjects.DataObjectType.App)
+            {
+                return NotFound();
+            }
+
             hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
 
             Models.DataObjectItem? DataObject = DataObjects.GetDataObject(ObjectType, Id);
@@ -254,6 +393,12 @@ namespace hasheous_server.Controllers.v1_0
         [Route("{ObjectType}/{Id}/SignatureMap/{SignatureMapId}")]
         public async Task<IActionResult> DeleteDataObjectSignatureMap(Classes.DataObjects.DataObjectType ObjectType, long Id, long SignatureId)
         {
+            // signatures aren't valid for apps
+            if (ObjectType != Classes.DataObjects.DataObjectType.App)
+            {
+                return NotFound();
+            }
+
             hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
 
             Models.DataObjectItem? DataObject = DataObjects.GetDataObject(ObjectType, Id);
@@ -277,6 +422,12 @@ namespace hasheous_server.Controllers.v1_0
         [Route("{ObjectType}/{Id}/MetadataMap")]
         public async Task<IActionResult> GetDataObjectMetadataMap(Classes.DataObjects.DataObjectType ObjectType, long Id)
         {
+            // metadata mappings aren't valid for apps
+            if (ObjectType != Classes.DataObjects.DataObjectType.App)
+            {
+                return NotFound();
+            }
+
             hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
 
             Models.DataObjectItem? DataObject = DataObjects.GetDataObject(ObjectType, Id);
@@ -298,6 +449,12 @@ namespace hasheous_server.Controllers.v1_0
         [Route("{ObjectType}/{Id}/SignatureSearch/")]
         public async Task<IActionResult> GetSignatureSearch(Classes.DataObjects.DataObjectType ObjectType, long Id, string SearchString)
         {
+            // signatures aren't valid for apps
+            if (ObjectType != Classes.DataObjects.DataObjectType.App)
+            {
+                return NotFound();
+            }
+
             hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
 
             Models.DataObjectItem? DataObject = DataObjects.GetDataObject(ObjectType, Id);
@@ -319,6 +476,12 @@ namespace hasheous_server.Controllers.v1_0
         [Route("{ObjectType}/{Id}/MergeObject/")]
         public async Task<IActionResult> MergeObjects(Classes.DataObjects.DataObjectType ObjectType, long Id, long TargetId, bool Commit = false)
         {
+            // merging isn't valid for apps
+            if (ObjectType != Classes.DataObjects.DataObjectType.App)
+            {
+                return NotFound();
+            }
+
             hasheous_server.Classes.DataObjects DataObjects = new Classes.DataObjects();
 
             Models.DataObjectItem? DataObject = DataObjects.GetDataObject(ObjectType, Id);
@@ -339,6 +502,80 @@ namespace hasheous_server.Controllers.v1_0
                 {
                     return Ok(DataObjects.MergeObjects(DataObject, TargetDataObject, Commit));
                 }
+            }
+        }
+
+        [MapToApiVersion("1.0")]
+        [HttpGet]
+        [Authorize]
+        [ProducesResponseType(typeof(List<ClientApiKeyItem>), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [Route("app/{Id}/ClientApiKeys")]
+        public async Task<IActionResult> GetClientApiKeys(long Id)
+        {
+            var user = await _userManager.GetUserAsync(User);
+
+            DataObjectPermission dataObjectPermission = new DataObjectPermission(_userManager);
+
+            if (dataObjectPermission.CheckAsync(user, DataObjects.DataObjectType.App, DataObjectPermission.PermissionType.Update, Id).Result)
+            {
+                Authentication.ClientApiKey clientApiKey = new Authentication.ClientApiKey();
+
+                return Ok(clientApiKey.GetApiKeys(Id));
+            }
+            else
+            {
+                return NotFound();
+            }
+        }
+
+        [MapToApiVersion("1.0")]
+        [HttpPost]
+        [Authorize]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [Route("app/{Id}/ClientApiKeys")]
+        public async Task<IActionResult> NewClientApiKey(long Id, string Name, DateTime? Expires)
+        {
+            var user = await _userManager.GetUserAsync(User);
+
+            DataObjectPermission dataObjectPermission = new DataObjectPermission(_userManager);
+
+            if (dataObjectPermission.CheckAsync(user, DataObjects.DataObjectType.App, DataObjectPermission.PermissionType.Update, Id).Result)
+            {
+                Authentication.ClientApiKey clientApiKey = new Authentication.ClientApiKey();
+
+                return Ok(clientApiKey.CreateApiKey(Id, Name, Expires));
+            }
+            else
+            {
+                return NotFound();
+            }
+        }
+
+        [MapToApiVersion("1.0")]
+        [HttpDelete]
+        [Authorize]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [Route("app/{Id}/ClientApiKeys/{ClientId}")]
+        public async Task<IActionResult> DeleteClientApiKey(long Id, long ClientId)
+        {
+            var user = await _userManager.GetUserAsync(User);
+
+            DataObjectPermission dataObjectPermission = new DataObjectPermission(_userManager);
+
+            if (dataObjectPermission.CheckAsync(user, DataObjects.DataObjectType.App, DataObjectPermission.PermissionType.Update, Id).Result)
+            {
+                Authentication.ClientApiKey clientApiKey = new Authentication.ClientApiKey();
+
+                clientApiKey.RevokeApiKey(Id, ClientId);
+
+                return Ok();
+            }
+            else
+            {
+                return NotFound();
             }
         }
     }
